requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
email-validator==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
tenacity==8.2.3
redis==5.0.1
hiredis==2.3.2
# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# LDAP para Federated Identity
ldap3==2.9.1